% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whopals.R
\name{pal_gender}
\alias{pal_gender}
\title{Gender / sex category colours}
\usage{
pal_gender(theme = c("light", "dark"))
}
\arguments{
\item{theme}{Either \code{"light"} or \code{"dark"}.}
}
\value{
Named vector \code{male}, \code{female}, \code{misc}.
}
\description{
Gender / sex category colours
}
