% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whopals.R
\name{pal_background}
\alias{pal_background}
\title{Background theme colours}
\usage{
pal_background(
  component = c("base", "weaker", "selection"),
  theme = c("light", "dark")
)
}
\arguments{
\item{component}{One of \code{base}, \code{weaker}, \code{selection}.}

\item{theme}{Either \code{"light"} or \code{"dark"}.}
}
\value{
Named character vector of length 1 (name = \code{component}).
}
\description{
Default \strong{\code{component}} is \strong{\code{base}}.
}
