% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whopals.R
\name{pal_text}
\alias{pal_text}
\title{Text theme colours}
\usage{
pal_text(component = c("base", "weaker"), theme = c("light", "dark"))
}
\arguments{
\item{component}{One of \code{base}, \code{weaker}.}

\item{theme}{Either \code{"light"} or \code{"dark"}.}
}
\value{
Named character vector of length 1 (name = \code{component}).
}
\description{
Default \strong{\code{component}} is \strong{\code{base}}.
}
